Change buttons to toggle buttons in changes window
The toggle buttons remain pressed-in when clicked, this with radio
action allows only one of the three to be pressed which provides visual
feedback on which type of view is currently selected. With normal
button it was impossible to figure out which view is currently selected
just by looking at the buttons.