Add StartupWMClass to the desktop file
This is necessary so that GNOME and others window managers can associate the
application window with the launcher icon. Without StartupWMClass, GNOME dock
would display another icon for the application when it’s launched instead of
adding a dot to show the app is running to the already existing one.